Move sql specific utils to shared location.
