+++ tests/integration/test_utils/data_source_config/sql.py
@@ -1,13 +1,16 @@
 from __future__ import annotations
 
+import logging
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
+from dataclasses import asdict, dataclass
 from datetime import date, datetime
 from functools import cached_property
-from typing import TYPE_CHECKING, Any, Callable, Generic, Mapping, Optional, Sequence, Union
+from typing import Any, Callable, Generic, Mapping, Optional, Sequence, Union
 
 import numpy as np
 import pandas as pd
+import sqlalchemy as sa
+from sqlalchemy.pool import QueuePool
 from typing_extensions import override
 
 from great_expectations.compatibility.sqlalchemy import (
@@ -23,11 +26,90 @@ from great_expectations.compatibility.sqlalchemy import (
 from great_expectations.data_context import AbstractDataContext
 from great_expectations.datasource.fluent.interfaces import Batch
 from great_expectations.datasource.fluent.sql_datasource import TableAsset
-from tests.integration.conftest import ConnectionDetails, TestSessionSQLEngineManager
 from tests.integration.test_utils.data_source_config.base import BatchTestSetup, _ConfigT
 
-if TYPE_CHECKING:
-    import sqlalchemy as sa
+logger = logging.getLogger(__name__)
+
+
+@dataclass(frozen=True)
+class ConnectionDetails:
+    connection_string: str
+    dialect: str
+
+
+@dataclass(frozen=True)
+class PoolConfig:
+    pool_size: int
+    max_overflow: int
+    pool_recycle: int
+    pool_timeout: int
+    pool_pre_ping: bool
+
+
+class TestSessionSQLEngineManager:
+    POOL_CONFIG = PoolConfig(
+        pool_size=2,
+        max_overflow=3,
+        pool_recycle=5400,  # 1.5 hours
+        pool_timeout=30,  # 30 seconds
+        pool_pre_ping=True,
+    )
+
+    def __init__(self):
+        self._engine_cache: dict[ConnectionDetails, sa.engine.Engine] = {}
+
+    def get_engine(
+        self,
+        connection_details: ConnectionDetails,
+    ) -> sa.engine.Engine:
+        cache_key = connection_details
+        if cache_key not in self._engine_cache:
+            logger.info(f"Cache miss for engine: {cache_key}. Creating new engine.")
+            engine_kwargs = asdict(self.POOL_CONFIG)
+            logger.info(
+                f"Creating engine for {connection_details.dialect} with settings: {engine_kwargs}"
+            )
+            self._engine_cache[cache_key] = sa.create_engine(
+                connection_details.connection_string, **engine_kwargs
+            )
+        else:
+            logger.info(f"Cache hit for engine: {cache_key}")
+        return self._engine_cache[cache_key]
+
+    def dispose_all_engines(self):
+        logger.info("Disposing all cached SQLAlchemy engines.")
+        for key, engine in self._engine_cache.items():
+            logger.info(f"Disposing engine: {key}")
+            try:
+                engine.dispose()
+            except Exception:
+                logger.exception(f"Error disposing engine '{key}'")
+        self._engine_cache.clear()
+
+    def get_all_pool_statistics(self) -> dict[ConnectionDetails, dict[str, Any]]:
+        stats = {}
+        for key, engine in self._engine_cache.items():
+            try:
+                pool = engine.pool
+                if isinstance(pool, QueuePool):
+                    stats[key] = {
+                        "size": pool.size(),
+                        "checked_in": pool.checkedin(),
+                        "overflow": pool.overflow(),
+                        "checked_out": pool.checkedout(),
+                    }
+                else:
+                    logger.warning(
+                        f"Pool for engine {key} is not a QueuePool. It is a {type(pool)}."
+                    )
+                    stats[key] = {
+                        "type": f"{type(pool)}",
+                        "status": f"{pool.status()}",
+                    }
+            except Exception as e:
+                logger.exception(f"Error getting pool status for engine '{key}'")
+                stats[key] = {"error": str(e)}
+        return stats
 
 
 @dataclass(frozen=True)
